maximize minimap use and disable freezecam
When the minimap is always centered on the player, it can waste
a lot of space. Either some of the map isn't shown or the view
must be zoomed out, making the whole map smaller and harder to
read.

These are more settings from Sapphire's recommended config for
event observers, which is now at
http://www.sapphireesports.com/single-post/2016/09/02/Setting-Config-Standards-for-Event-Observers