do not rotate map
The map does not rotate in larger views, so this will be more
consistent. With the fixed map, it should also be easier to know
exactly where things happen after a while, learning locations.

The benefit of rotating map is that it's easier to use on unfamiliar
maps for nearby things.