Revert "Switch to ESDF from WASD"
This reverts commit c71734f62d277aba4b4d2a9a3d02c3c9f3bfc78c.

While ESDF would be better than WASD to avoid Alt+Tab, OS key,
and ~ for the dev console, the support for remapping is poor
enough in some games that ESDF isn't a realistic option. Many
maps and menus, for example, are hard-coded for WASD.

For consistency, revert to WASD for all games.